Fix non working links
Some links were unintentionally concatenating by adding the root website to the link. Found out that when it was uploaded that the quotation marks were on different format. I used Notepad++ to see the difference. Updated font format to make it more readable. Also added favicon for better identification on browser tabs.